LayoutTests/js/dom/resizable-array-buffer-view-serialization-out-of-bounds-expected.txt

Resizable ArrayBuffers should be serializable OOB

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS arrayBuffer.resizable is true
PASS arrayBuffer.byteLength is 36
PASS arrayBuffer.maxByteLength is 128
PASS int32AutoArray.length is 8
PASS int32AutoArray.byteOffset is 4
PASS structuredClone(int32AutoArray) threw exception DataCloneError: The object can not be cloned..
PASS arrayBuffer.resizable is true
PASS arrayBuffer.byteLength is 0
PASS arrayBuffer.maxByteLength is 128
PASS int32AutoArray.length is 0
PASS int32AutoArray.byteOffset is 0
PASS arrayBuffer.resizable is true
PASS arrayBuffer.byteLength is 36
PASS arrayBuffer.maxByteLength is 128
PASS int32AutoArray.length is 8
PASS int32AutoArray.byteOffset is 4
PASS successfullyParsed is true

TEST COMPLETE
